+++ tests/data/test_update/v3/addon_d2/__openerp__.py
@@ -1,0 +1,1 @@
+{"name": "addon_d2", "version": "3.0"}