+++ tests/data/test_update/v4/addon_d1/__openerp__.py
@@ -1,0 +1,1 @@
+{"name": "addon_d1", "version": "4.0"}