+++ tests/data/test_update/v5/addon_app/__openerp__.py
@@ -1,0 +1,7 @@
+{
+    "name": "addon_app",
+    "application": True,
+    "version": "5.0",
+    "depends": ["addon_d1"],
+    "description": "Uninstall addon_d2 with a migration.",
+}