+++ tests/data/test_update/v5/addon_app/migrations/5.0/pre-migrate.py
@@ -1,0 +1,2 @@
+def migrate(cr, version):
+    cr.execute("UPDATE ir_module_module SET state='to remove' WHERE name='addon_d2'")