+++ tests/data/test_update/v5/addon_d1/__openerp__.py
@@ -1,0 +1,1 @@
+{"name": "addon_d1", "version": "5.0"}