+++ CHANGES.rst
@@ -5,6 +5,15 @@ Changes
 .. ----------
 .. -
 
+1.4.0 (2018-11-19)
+------------------
+
+- new click-odoo-update which implements the functionality of module_auto_update
+  natively, alleviating the need to have module_auto_update installed in the database,
+  and is more robust (it does a regular -u after identifying modules to update)
+- upgrade: deprecated in favor of click-odoo-update
+- initdb: save installed checksums so click-odoo-update can readily use them
+- initdb: add --addons-path option
 - copydb: fix error when source filestore did not exist
 
 1.3.1 (2018-11-05)